master: move out db
